Clean up css, create LeftSidebar shell
