+++ src/assets/css/prem-chat.css
@@ -8,17 +8,11 @@
 }
 
 .prem-chat-sidebar .sidebar {
-  @apply max-md:fixed max-md:z-[11] max-md:w-full max-md:p-5 max-md:overflow-hidden;
+  @apply max-md:fixed max-md:z-[11] max-md:w-full max-md:p-5 max-md:top-0 max-md:bottom-0 max-md:overflow-hidden max-md:bg-grey-800;
 }
 
-@media (max-width: 767px) {
-  .prem-chat-sidebar .sidebar {
-    @apply top-0 bottom-0 overflow-hidden bg-grey-800;
-  }
-
-  .prem-chat-sidebar .sidebar li.active {
-    @apply bg-[#1A1E23];
-  }
+.prem-chat-sidebar .sidebar li.active {
+  @apply max-md:bg-grey-900;
 }
 
 .prem-chat-sidebar .sidebar__logo {
