+++ src/modules/prem-audio/components/PremAudioLeftSidebar.tsx
@@ -1,10 +1,8 @@
-import leftArrow from "assets/images/arrow.svg";
-import cross from "assets/images/cross.svg";
-import Logo from "assets/images/logo.svg";
 import clsx from "clsx";
 import { orderBy } from "lodash";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import DeleteIcon from "shared/components/DeleteIcon";
+import LeftSidebar from "shared/components/LeftSidebar";
 import NoPrompts from "shared/components/NoPrompts";
 import usePremAudioStore from "shared/store/prem-audio";
 import type { HamburgerMenuProps } from "shared/types";
@@ -23,10 +21,6 @@ const PremAudioLeftSidebar = ({ setHamburgerMenu }: HamburgerMenuProps) => {
     shallow,
   );
 
-  const onCloseClick = () => {
-    navigate("/");
-  };
-
   const onDeleteClick = (id: string) => {
     deleteHistory(id);
     if (historyId === id) {
@@ -35,44 +29,29 @@ const PremAudioLeftSidebar = ({ setHamburgerMenu }: HamburgerMenuProps) => {
   };
 
   return (
-    <>
-      <div className="md:pt-7 !bg-grey-800 pt-[22px] pb-[10px] flex-col px-2 flex md:h-screen sidebar md:!w-[259px]">
-        <div className="flex max-md:flex-wrap md:gap-4">
-          <button className="mx-2" onClick={onCloseClick}>
-            <img src={leftArrow} alt="left-arrow" className="max-md:max-w-[15px]" />
-          </button>
-          <img className="sidebar__logo" src={Logo} alt="logo" />
-          <button
-            onClick={() => setHamburgerMenu(true)}
-            className="w-[30px] md:hidden z-[11] fixed right-[14px] top-[24px]"
-          >
-            <img src={cross} alt="cross" width={22} height={22} className="mx-auto" />
-          </button>
-        </div>
-
-        {history.length === 0 && <NoPrompts text="No Audio" />}
-        <div className="mt-10 overflow-y-auto custom-scroll">
-          <ul className="md:flex-grow scrollbar-none w-full">
-            {orderBy(history, "timestamp", "desc").map((item) => {
-              return (
-                <li
-                  onClick={() => responsiveMatches && setHamburgerMenu(true)}
-                  key={item.id}
-                  className={clsx({ "md:bg-grey-900 bg-[#1A1E23]": historyId === item.id })}
-                >
-                  <Link to={`/prem-audio/${serviceId}/${item.id}`}>
-                    <span className="text-white">{item.file}</span>
-                  </Link>
-                  <button onClick={() => onDeleteClick(item.id)}>
-                    <DeleteIcon />
-                  </button>
-                </li>
-              );
-            })}
-          </ul>
-        </div>
+    <LeftSidebar setHamburgerMenu={setHamburgerMenu}>
+      {history.length === 0 && <NoPrompts text="No Audio" />}
+      <div className="mt-10 overflow-y-auto custom-scroll">
+        <ul className="md:flex-grow scrollbar-none w-full">
+          {orderBy(history, "timestamp", "desc").map((item) => {
+            return (
+              <li
+                onClick={() => responsiveMatches && setHamburgerMenu(true)}
+                key={item.id}
+                className={clsx({ "bg-grey-900": historyId === item.id })}
+              >
+                <Link to={`/prem-audio/${serviceId}/${item.id}`}>
+                  <span className="text-white">{item.file}</span>
+                </Link>
+                <button onClick={() => onDeleteClick(item.id)}>
+                  <DeleteIcon />
+                </button>
+              </li>
+            );
+          })}
+        </ul>
       </div>
-    </>
+    </LeftSidebar>
   );
 };
