+++ src/modules/prem-image/components/PremImageLeftSidebar.tsx
@@ -1,18 +1,16 @@
-import leftArrow from "assets/images/arrow.svg";
-import cross from "assets/images/cross.svg";
-import Logo from "assets/images/logo.svg";
 import clsx from "clsx";
 import { format, parseISO } from "date-fns";
 import { orderBy } from "lodash";
-import { Link, useNavigate, useParams } from "react-router-dom";
+import { Link, useParams } from "react-router-dom";
 import NoPrompts from "shared/components/NoPrompts";
 import usePremImageStore from "shared/store/prem-image";
-import type { HamburgerMenuProps } from "shared/types";
 import { useMediaQuery } from "usehooks-ts";
 import { shallow } from "zustand/shallow";
 
+import LeftSidebar from "../../../shared/components/LeftSidebar";
+import type { HamburgerMenuProps } from "../../../shared/types";
+
 const PremImageLeftSidebar = ({ setHamburgerMenu }: HamburgerMenuProps) => {
-  const navigate = useNavigate();
   const { serviceId, historyId } = useParams();
   const responsiveMatches = useMediaQuery("(max-width: 767px)");
   const { history } = usePremImageStore(
@@ -22,10 +20,6 @@ const PremImageLeftSidebar = ({ setHamburgerMenu }: HamburgerMenuProps) => {
     shallow,
   );
 
-  const onCloseClick = () => {
-    navigate("/");
-  };
-
   const scrollToTop = () => {
     responsiveMatches && setHamburgerMenu(true);
     document
@@ -34,54 +28,40 @@ const PremImageLeftSidebar = ({ setHamburgerMenu }: HamburgerMenuProps) => {
   };
 
   return (
-    <>
-      <div className="md:pt-7 !bg-grey-800 pt-[22px] pb-[10px] flex-col px-2 flex md:h-screen sidebar md:!w-[259px]">
-        <div className="flex max-md:flex-wrap md:gap-4">
-          <button className="mx-2" onClick={onCloseClick}>
-            <img src={leftArrow} alt="left-arrow" className="max-md:max-w-[15px]" />
-          </button>
-          <img className="sidebar__logo" src={Logo} alt="logo" />
-          <button
-            onClick={() => setHamburgerMenu(true)}
-            className="w-[30px] md:hidden z-[11] fixed right-[14px] top-[24px]"
-          >
-            <img src={cross} alt="cross" width={22} height={22} className="mx-auto" />
-          </button>
-        </div>
-        {history.length === 0 && <NoPrompts text="No Image" />}
-        <div className="prem-image-sidebar custom-scroll">
-          <ul className="md:flex-grow scrollbar-none w-full">
-            {orderBy(history, "timestamp", "desc").map((item) => {
-              return (
-                <li
-                  onClick={scrollToTop}
-                  key={item.id}
-                  className={clsx({ "md:bg-grey-900 bg-[#1A1E23]": historyId === item.id })}
-                >
-                  <Link to={`/prem-image/${serviceId}/${item.id}`}>
-                    <div className="flex w-full">
-                      <div>
-                        <span className="text-white">
-                          {format(parseISO(item.timestamp), "LLLL dd, hh:mm a")}
-                        </span>
-                        <div className="flex flex-wrap gap-[2px] mt-[11px]">
-                          {item.images?.map((image, index) => {
-                            return <img key={index} src={image} className="w-8 h-8" alt="" />;
-                          })}
-                        </div>
-                      </div>
-                      <span className="text-[#58595E] text-end">
-                        {item.images.length}&nbsp;images
+    <LeftSidebar setHamburgerMenu={setHamburgerMenu}>
+      {history.length === 0 && <NoPrompts text="No Image" />}
+      <div className="prem-image-sidebar custom-scroll">
+        <ul className="md:flex-grow scrollbar-none w-full">
+          {orderBy(history, "timestamp", "desc").map((item) => {
+            return (
+              <li
+                onClick={scrollToTop}
+                key={item.id}
+                className={clsx({ "bg-grey-900": historyId === item.id })}
+              >
+                <Link to={`/prem-image/${serviceId}/${item.id}`}>
+                  <div className="flex w-full">
+                    <div>
+                      <span className="text-white">
+                        {format(parseISO(item.timestamp), "LLLL dd, hh:mm a")}
                       </span>
+                      <div className="flex flex-wrap gap-[2px] mt-[11px]">
+                        {item.images?.map((image, index) => {
+                          return <img key={index} src={image} className="w-8 h-8" alt="" />;
+                        })}
+                      </div>
                     </div>
-                  </Link>
-                </li>
-              );
-            })}
-          </ul>
-        </div>
+                    <span className="text-[#58595E] text-end">
+                      {item.images.length}&nbsp;images
+                    </span>
+                  </div>
+                </Link>
+              </li>
+            );
+          })}
+        </ul>
       </div>
-    </>
+    </LeftSidebar>
   );
 };
